AoC 2020 Day 25: Improve type safety in calcLoopSize
It returns an int64_t, so the variable ("loopSize") it returns should be of that
type, too, rather than "int" like it was before!

Doesn't actually make any difference in this case (Fermat's Little
Theorem tells us loopSize will never exceed 20201227, which fits
comfortably into an int), but let's not be sloppy! :p

+++ adventofcode/2020/day25/day25_part1_solution.cpp
@@ -9,7 +9,7 @@ int64_t calcLoopSize(int64_t publicKey)
 {
     constexpr int subjectNumber = 7;
     int64_t value = 1;
-    int loopSize = 0;
+    int64_t loopSize = 0;
     while (value != publicKey)
     {
         value = (value * subjectNumber) % Mod;
